+++ frontend/src/main.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import './App.css';
-import Navbar from './components/Navbar'
+import Navbar from './components/Navbar/Navbar.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <Navbar/>
+    <Navbar />
     <App />
   </React.StrictMode>
 );
